backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic>=2.9,<2.12
pydantic-settings==2.10.1

motor==3.3.2
pymongo==4.8.0

python-dotenv==1.0.0
zhipuai==2.0.1
jinja2==3.1.3

# LangChain 组合 + 强制固定 core 版本避免被解到 1.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79

# Py3.13 对应官方 wheel
numpy==2.3.4

pyyaml==6.0.2
httpx==0.26.0

# 测试栈（与 Py3.13 兼容）
pytest==8.4.2
pytest-asyncio==0.24.0
